Title: 합과 차의 나눗셈, Time: 108ms, Memory: 24MB, Status: Wrong Answer - CodeTree

+++ 241026/합과 차의 나눗셈/divide-of-sum-and-sub.py
@@ -1,4 +1,4 @@
 import sys
 input = sys.stdin.readline
-a,b = list(map(int,input().split()))
+a,b = list(map(float,input().split()))
 print(round((a+b)/(a-b),2))
